+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.2.2
+
++ Added support for platyPS and initial markdown documents
++ Added one touch deploy functions
+
 ## 0.2.1
 
 + Completed build script